// src/types/manga.ts
export interface Manga {
  _id: string;
  account_id: string; // Reference to User (_id)
  group_id?: string; // Reference to Group (_id) - Optional
  title: string;
  alternative_title?: string[]; // Optional
  language?: string; // Reference to Language (_id) - Optional
  author: string;
  artist?: string; // Optional
  description: string;
  manga_cover: string;
  viewCount: number;
  genre?: string[]; // Array of Genre (_id) - Optional
  status: 'ongoing' | 'hiatus' | 'completed' | 'draft' | 'deleted';
  createdAt?: string; // Auto-generated by Mongoose (Optional)
  updatedAt?: string; // Auto-generated by Mongoose (Optional)
}

export interface ReadingHistory {
  mangaId: string;
  title: string;
  manga_cover: string;
  lastReadChapter: number;
  lastReadTime: string;
}